refactor: move exceptions from api layer to service layer and rename constants to utils
