+++ dukan/src/main/kotlin/net/thechance/dukan/api/utils/EndPoints.kt
@@ -1,4 +1,4 @@
-package net.thechance.dukan.api.constants
+package net.thechance.dukan.api.utils
 
 internal object EndPoints {
     const val DUKAN_PATH ="/dukan"
