+++ dukan/src/main/kotlin/net/thechance/dukan/service/DukanProductService.kt
@@ -3,9 +3,9 @@ package net.thechance.dukan.service
 import jakarta.persistence.EntityNotFoundException
 import jakarta.transaction.Transactional
 import net.thechance.dukan.entity.DukanProduct
-import net.thechance.dukan.api.exception.DukanProductCreationFailedException
-import net.thechance.dukan.api.exception.ProductNameAlreadyTakenException
-import net.thechance.dukan.api.exception.ProductNotFoundException
+import net.thechance.dukan.service.exception.DukanProductCreationFailedException
+import net.thechance.dukan.service.exception.ProductNameAlreadyTakenException
+import net.thechance.dukan.service.exception.ProductNotFoundException
 import net.thechance.dukan.repository.DukanProductRepository
 import net.thechance.dukan.repository.DukanShelfRepository
 import net.thechance.dukan.service.model.DukanProductCreationParams
