+++ dukan/src/main/kotlin/net/thechance/dukan/service/DukanService.kt
@@ -5,8 +5,8 @@ import jakarta.transaction.Transactional
 import net.thechance.dukan.entity.Dukan
 import net.thechance.dukan.entity.DukanCategory
 import net.thechance.dukan.entity.DukanColor
-import net.thechance.dukan.api.exception.DukanCreationFailedException
-import net.thechance.dukan.api.exception.DukanNotFoundException
+import net.thechance.dukan.service.exception.DukanCreationFailedException
+import net.thechance.dukan.service.exception.DukanNotFoundException
 import net.thechance.dukan.api.mapper.dukan.toDukan
 import net.thechance.dukan.repository.DukanCategoryRepository
 import net.thechance.dukan.repository.DukanColorRepository
