+++ dukan/src/main/kotlin/net/thechance/dukan/service/exception/DukanProductExceptions.kt
@@ -1,8 +1,8 @@
-package net.thechance.dukan.api.exception
+package net.thechance.dukan.service.exception
 
-import net.thechance.dukan.api.constants.ErrorCodes.DUKAN_PRODUCT_CREATION_FAILED
-import net.thechance.dukan.api.constants.ErrorCodes.PRODUCT_NAME_ALREADY_TAKEN
-import net.thechance.dukan.api.constants.ErrorCodes.PRODUCT_NOT_FOUND
+import net.thechance.dukan.api.utils.ErrorCodes.DUKAN_PRODUCT_CREATION_FAILED
+import net.thechance.dukan.api.utils.ErrorCodes.PRODUCT_NAME_ALREADY_TAKEN
+import net.thechance.dukan.api.utils.ErrorCodes.PRODUCT_NOT_FOUND
 import org.springframework.http.HttpStatus
 
 class ProductNotFoundException() : DukanException(
