+++ dukan/src/main/kotlin/net/thechance/dukan/service/exception/DukanShelfExceptions.kt
@@ -1,8 +1,8 @@
-package net.thechance.dukan.api.exception
+package net.thechance.dukan.service.exception
 
-import net.thechance.dukan.api.constants.ErrorCodes.SHELF_DELETION_NOT_ALLOWED
-import net.thechance.dukan.api.constants.ErrorCodes.SHELF_NAME_ALREADY_TAKEN
-import net.thechance.dukan.api.constants.ErrorCodes.SHELF_NOT_FOUND
+import net.thechance.dukan.api.utils.ErrorCodes.SHELF_DELETION_NOT_ALLOWED
+import net.thechance.dukan.api.utils.ErrorCodes.SHELF_NAME_ALREADY_TAKEN
+import net.thechance.dukan.api.utils.ErrorCodes.SHELF_NOT_FOUND
 import org.springframework.http.HttpStatus
